// src/shared/dataprovider/api/model/transactionSummaryIncomeTotal.ts
/**
 * DynastyGateway API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface TransactionSummaryIncomeTotal { 
    IncomeToday?: number;
    NetIncomeToday?: number;
    IncomeYesterday?: number;
    NetIncomeYesterday?: number;
    IncomeThisWeek?: number;
    NetIncomeThisWeek?: number;
    IncomeThisMonth?: number;
    NetIncomeThisMonth?: number;
    Income?: number;
    NetIncome?: number;
}